develop: Add tags and move controllers to actions
