+++ backend/composer.json
@@ -28,7 +28,8 @@
         "symfony/templating": "4.3.*",
         "symfony/translation": "4.3.*",
         "symfony/validator": "4.3.*",
-        "symfony/yaml": "4.3.*"
+        "symfony/yaml": "4.3.*",
+        "ext-http": "*"
     },
     "require-dev": {
         "doctrine/doctrine-fixtures-bundle": "^3.2",
